backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.30
asyncpg>=0.30.0
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
stripe>=9.9.0
python-dotenv>=1.0.1
slowapi>=0.1.9
starlette>=0.40.0,<0.42.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
